refactor: global/auth의 entity,repository,controller위치 변경
